Move create_spec shortcut
